Shared DisaggregationMode enum for vLLM and SGLang, rename prefill_and_decode to agg
