+++ components/src/dynamo/common/__init__.py
@@ -12,7 +12,7 @@ Main submodules:
     - utils: Common utilities including environment and prometheus helpers
 """
 
-from dynamo.common import config_dump, utils
+from dynamo.common import config_dump, constants, utils
 
 try:
     from ._version import __version__
@@ -24,4 +24,4 @@ except Exception:
     except Exception:
         __version__ = "0.0.0+unknown"
 
-__all__ = ["__version__", "config_dump", "utils"]
+__all__ = ["__version__", "config_dump", "constants", "utils"]
