+++ components/src/dynamo/vllm/tests/test_vllm_unit.py
@@ -292,12 +292,12 @@ def test_conflicting_legacy_and_new_flags_raises(mock_vllm_cli):
 
 
 def test_explicit_default_mode_with_legacy_flag_raises(mock_vllm_cli):
-    """Test that --disaggregation-mode prefill_and_decode --is-decode-worker raises ValueError."""
+    """Test that --disaggregation-mode agg --is-decode-worker raises ValueError."""
     mock_vllm_cli(
         "--model",
         "Qwen/Qwen3-0.6B",
         "--disaggregation-mode",
-        "prefill_and_decode",
+        "agg",
         "--is-decode-worker",
     )
     with pytest.raises(ValueError, match="Cannot combine"):
